# docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Invariant Fabric engineering API",
  "version": "1.0.0",
  "description": "Exact-action software enforcement and synthetic target execution. Not a production-certified deployment. JSON is restricted to IF-CJSON-1. Cookie mutations require Origin and X-CSRF-Token; bearer credentials are also supported. Strict runtime validation is authoritative."
 },
 "servers": [
  {
   "url": "http://127.0.0.1:8080"
  }
 ],
 "security": [
  {
   "bearerAuth": []
  }
 ],
 "paths": {
  "/v1/me": {
   "get": {
    "operationId": "get_v1_me",
    "summary": "Current principal",
    "description": "Roles: authenticated. Engineering profile; all target mutations are simulated.",
    "tags": [
     "Control"
    ],
    "responses": {
     "200": {
      "description": "Successful response; see API.md for exact record contracts",
      "content": {
       "application/json": {
        "schema": {
         "type": "object"
        }
       }
      }
     },
     "default": {
      "description": "Reason-coded rejection",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    }
   }
  },
  "/v1/schemas": {
   "get": {
    "operationId": "get_v1_schemas",
    "summary": "Typed schema definitions and digests",
    "description": "Roles: authenticated. Engineering profile; all target mutations are simulated.",
    "tags": [
     "Control"
    ],
    "responses": {
     "200": {
      "description": "Successful response; see API.md for exact record contracts",
      "content": {
       "application/json": {
        "schema": {
         "type": "object"
        }
       }
      }
     },
     "default": {
      "description": "Reason-coded rejection",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    }
   }
  },
  "/v1/policy": {
   "get": {
    "operationId": "get_v1_policy",
    "summary": "Read active constitution",
    "description": "Roles: operator, approver, custodian, policy_admin, security. Engineering profile; all target mutations are simulated.",
    "tags": [
     "Control"
    ],
    "responses": {
     "200": {
      "description": "Successful response; see API.md for exact record contracts",
      "content": {
       "application/json": {
        "schema": {
         "type": "object"
        }
       }
      }
     },
     "default": {
      "description": "Reason-coded rejection",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    }
   }
  },
  "/v1/action-capsules": {
   "get": {
    "operationId": "get_v1_action_capsules",
    "summary": "List capsules (limit 1\u2013100 and offset)",
    "description": "Roles: operator, approver, custodian, security, policy_admin. Engineering profile; all target mutations are simulated.",
    "tags": [
     "Control"
    ],
    "responses": {
     "200": {
      "description": "Successful response; see API.md for exact record contracts",
      "content": {
       "application/json": {
        "schema": {
         "type": "object"
        }
       }
      }
     },
     "default": {
      "description": "Reason-coded rejection",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    }
   },
   "post": {
    "operationId": "post_v1_action_capsules",
    "summary": "Propose exact action",
    "description": "Roles: operator, policy_admin, workload. Engineering profile; all target mutations are simulated.",
    "tags": [
     "Control"
    ],
    "responses": {
     "201": {
      "description": "Successful response; see API.md for exact record contracts",
      "content": {
       "application/json": {
        "schema": {
         "type": "object"
        }
       }
      }
     },
     "default": {
      "description": "Reason-coded rejection",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/Proposal"
       }
      }
     }
    },
    "parameters": [
     {
      "name": "Idempotency-Key",
      "in": "header",
      "required": true,
      "schema": {
       "type": "string",
       "pattern": "^[A-Za-z0-9_-]{8,128}$"
      }
     }
    ]
   }
  },
...
  "/v1/approvals": {
   "post": {
    "operationId": "post_v1_approvals",
    "summary": "Submit offline-signed approval",
    "description": "Roles: approver, custodian. Engineering profile; all target mutations are simulated.",
    "tags": [
     "Control"
    ],
    "responses": {
     "201": {
      "description": "Successful response; see API.md for exact record contracts",
      "content": {
       "application/json": {
        "schema": {
         "type": "object"
        }
       }
      }
     },
     "default": {
      "description": "Reason-coded rejection",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/Envelope"
       }
      }
     }
    }
   }
  },
  "/v1/certificates": {
   "post": {
    "operationId": "post_v1_certificates",
    "summary": "Mint single-use authority after ALLOW",
    "description": "Roles: operator, policy_admin. Engineering profile; all target mutations are simulated.",
    "tags": [
     "Control"
    ],
    "responses": {
     "201": {
      "description": "Successful response; see API.md for exact record contracts",
      "content": {
       "application/json": {
        "schema": {
         "type": "object"
        }
       }
      }
     },
     "default": {
      "description": "Reason-coded rejection",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CertificateRequest"
       }
      }
     }
    }
   }
  },
...
  "/gate/v1/execute": {
   "post": {
    "operationId": "post_gate_v1_execute",
    "summary": "Verify and execute or dry-run simulated target",
    "description": "Roles: operator, policy_admin. Engineering profile; all target mutations are simulated.",
    "tags": [
     "Gate"
    ],
    "responses": {
     "200": {
      "description": "Successful response; see API.md for exact record contracts",
      "content": {
       "application/json": {
        "schema": {
         "type": "object"
        }
       }
      }
     },
     "default": {
      "description": "Reason-coded rejection",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ExecuteRequest"
       }
      }
     }
    }
   }
  },
...
  "/v1/capabilities": {
   "post": {
    "operationId": "post_v1_capabilities",
    "summary": "Issue narrow local capability",
    "description": "Roles: operator, workload. Engineering profile; all target mutations are simulated.",
    "tags": [
     "Control"
    ],
    "responses": {
     "201": {
      "description": "Successful response; see API.md for exact record contracts",
      "content": {
       "application/json": {
        "schema": {
         "type": "object"
        }
       }
      }
     },
     "default": {
      "description": "Reason-coded rejection",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CapabilityRequest"
       }
      }
     }
    }
   }
  },
  "/gate/v1/runtime": {
   "post": {
    "operationId": "post_gate_v1_runtime",
    "summary": "Consume capability under shared budgets",
    "description": "Roles: bound subject. Engineering profile; all target mutations are simulated.",
    "tags": [
     "Gate"
    ],
    "responses": {
     "200": {
      "description": "Successful response; see API.md for exact record contracts",
      "content": {
       "application/json": {
        "schema": {
         "type": "object"
        }
       }
      }
     },
     "default": {
      "description": "Reason-coded rejection",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RuntimeRequest"
       }
      }
     }
    }
   }
  },
  "/v1/revocations": {
   "post": {
    "operationId": "post_v1_revocations",
    "summary": "Permanently revoke local authority",
    "description": "Roles: security. Engineering profile; all target mutations are simulated.",
    "tags": [
     "Control"
    ],
    "responses": {
     "201": {
      "description": "Successful response; see API.md for exact record contracts",
      "content": {
       "application/json": {
        "schema": {
         "type": "object"
        }
       }
      }
     },
     "default": {
      "description": "Reason-coded rejection",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/Revocation"
       }
      }
     }
    }
   }
  },
  "/v1/coverage": {
   "get": {
    "operationId": "get_v1_coverage",
    "summary": "Read signed conservative coverage manifest",
    "description": "Roles: operator, approver, custodian, security, auditor, policy_admin. Engineering profile; all target mutations are simulated.",
    "tags": [
     "Control"
    ],
    "responses": {
     "200": {
      "description": "Successful response; see API.md for exact record contracts",
      "content": {
       "application/json": {
        "schema": {
         "type": "object"
        }
       }
      }
     },
     "default": {
      "description": "Reason-coded rejection",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    }
   },
   "post": {
    "operationId": "post_v1_coverage",
    "summary": "Declare monitored or unknown path",
    "description": "Roles: security. Engineering profile; all target mutations are simulated.",
    "tags": [
     "Control"
    ],
    "responses": {
     "201": {
      "description": "Successful response; see API.md for exact record contracts",
      "content": {
       "application/json": {
        "schema": {
         "type": "object"
        }
       }
      }
     },
     "default": {
      "description": "Reason-coded rejection",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CoverageDeclaration"
       }
      }
     }
    }
   }
  },
  "/v1/connectors": {
   "get": {
    "operationId": "get_v1_connectors",
    "summary": "Read simulator connector limitations",
    "description": "Roles: authenticated. Engineering profile; all target mutations are simulated.",
    "tags": [
     "Control"
    ],
    "responses": {
     "200": {
      "description": "Successful response; see API.md for exact record contracts",
      "content": {
       "application/json": {
        "schema": {
         "type": "object"
        }
       }
      }
     },
     "default": {
      "description": "Reason-coded rejection",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    }
   }
  },
  "/v1/policies/simulate": {
   "post": {
    "operationId": "post_v1_policies_simulate",
    "summary": "Compare exact candidate without activation",
    "description": "Roles: policy_admin, security. Engineering profile; all target mutations are simulated.",
    "tags": [
     "Control"
    ],
    "responses": {
     "200": {
      "description": "Successful response; see API.md for exact record contracts",
      "content": {
       "application/json": {
        "schema": {
         "type": "object"
        }
       }
      }
     },
     "default": {
      "description": "Reason-coded rejection",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/Policy"
       }
      }
     }
    }
   }
  },
  "/v1/audit-exports": {
   "post": {
    "operationId": "post_v1_audit_exports",
    "summary": "Export audited purpose-bound integrity metadata",
    "description": "Roles: auditor, security. Engineering profile; all target mutations are simulated.",
    "tags": [
     "Control"
    ],
    "responses": {
     "200": {
      "description": "Successful response; see API.md for exact record contracts",
      "content": {
       "application/json": {
        "schema": {
         "type": "object"
        }
       }
      }
     },
     "default": {
      "description": "Reason-coded rejection",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AuditRequest"
       }
      }
     }
    }
   }
  },
  "/v1/retention/hold": {
   "post": {
    "operationId": "post_v1_retention_hold",
    "summary": "Set or release evidence legal hold",
    "description": "Roles: security. Engineering profile; all target mutations are simulated.",
    "tags": [
     "Control"
    ],
    "responses": {
     "200": {
      "description": "Successful response; see API.md for exact record contracts",
      "content": {
       "application/json": {
        "schema": {
         "type": "object"
        }
       }
      }
     },
     "default": {
      "description": "Reason-coded rejection",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RetentionHold"
       }
      }
     }
    }
   }
  },
  "/v1/retention/sweep": {
   "post": {
    "operationId": "post_v1_retention_sweep",
    "summary": "Apply conservative logical retention",
    "description": "Roles: security. Engineering profile; all target mutations are simulated.",
    "tags": [
     "Control"
    ],
    "responses": {
     "200": {
      "description": "Successful response; see API.md for exact record contracts",
      "content": {
       "application/json": {
        "schema": {
         "type": "object"
        }
       }
      }
     },
     "default": {
      "description": "Reason-coded rejection",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/Empty"
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "securitySchemes": {
   "bearerAuth": {
    "type": "http",
    "scheme": "bearer"
   }
  },
  "schemas": {
   "Error": {
    "type": "object",
    "additionalProperties": false,
    "required": [
     "error"
    ],
    "properties": {
     "error": {
      "type": "object",
      "additionalProperties": false,
      "required": [
       "code",
       "message",
       "request_id"
      ],
      "properties": {
       "code": {
        "type": "string"
       },
       "message": {
        "type": "string"
       },
       "request_id": {
        "type": "string"
       }
      }
     }
    }
   },
   "Envelope": {
    "type": "object",
    "additionalProperties": false,
    "required": [
     "protected",
     "payload",
     "signature"
    ],
    "properties": {
     "protected": {
      "type": "object",
      "additionalProperties": false,
      "required": [
       "profile",
       "suite",
       "key_id",
       "purpose"
      ],
      "properties": {
       "profile": {
        "const": "IF-CJSON-1"
       },
       "suite": {
        "const": "Ed25519"
       },
       "key_id": {
        "type": "string"
       },
       "purpose": {
        "type": "string"
       }
      }
     },
     "payload": {
      "type": "object"
     },
     "signature": {
      "type": "string",
      "pattern": "^[A-Za-z0-9_-]{86}$"
     }
    }
   },
...
  }
 }
}